+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Moved HTTP client specific code from `RestTemplateExchangeClient` to `CloseableHttpRequestFactory`.
 - Added fallback to Java net HTTP client if Apache HTTP Client 5 library is not present in the classpath in `RestTemplateExchangeClient`.
 - Added follow redirects property `ClientProperties.Connection` to configure redirect behavior.
+- Renamed `SpringHttpRequests` to `SpringHttpSupport`.
 - Upgraded `oauth2-oidc-sdk` to `11.37.1`.
 - Upgraded `org.eclipse.jdt.core` to `3.45.0`.
 - Upgraded `spotless-maven-plugin` to `3.4.0`.